[irods/irods#4005] add deprecation notice for isPathSymlink()
The API function isPathSymlink() is now marked deprecated.

isPathSymlink() returns 0 or 1 only, and when errors are encountered,
it returns 1, signifying that the parameter file should not be
included in operations.

The new function isPathSymlink_err() is provided instead, and it is able
to return the traditional irods return codes in addition to 0 and 1 for
paths that have no errors.

The new function has a slightly different "footprint" which provides
identical functionality, but also adds a negative return code for
"fatal errors" (such as dangling symbolic links, inaccessible files, etc).

The new return codes:

  0  - treat the parameter file as **not** a symlink
  1  - treat the parameter file as a symlink
 <0  - Error code (message in the message stack)