[irods/irods#7468] Trim: Minimum replicas includes good and stale
The definition of trim has been altered slightly to mean that the number
of replicas is reduced to some minimum positive number. This has some
implications which are now reflected in the documentation as well as an
explanation for the priority of replicas to trim. A more expanded table
demonstrating the selection process has been included.