[irods/irods#6940] Add documentation for unixfilesystem voting
The process for how storage resources vote has not been documented
outside of the code. This commit includes the voting values for all
the various situations in which a resource or a replica on the resource
may find themselves. These votes are, of course, filtered up to the
parent resource for decision making higher up, but these policies have
been documented for some time.